% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_tracer.R
\name{new}
\alias{new}
\title{new application}
\usage{
new(...)
}
\arguments{
\item{...}{No argument are needed to launch the application.}
}
\description{
An application to perform hydrological modelling of tracers.
}
