% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{RaDisMod}
\alias{RaDisMod}
\title{RaDisMod application}
\usage{
RaDisMod(...)
}
\arguments{
\item{...}{No argument are needed to launch the application.}
}
\description{
An application to perform hydrological modelling of single flood events.
}
